+++ src/node/interfaces.cpp
@@ -4,7 +4,6 @@
 
 #include <addrdb.h>
 #include <banman.h>
-#include <boost/signals2/signal.hpp>
 #include <chain.h>
 #include <chainparams.h>
 #include <init.h>
@@ -17,6 +16,7 @@
 #include <net_processing.h>
 #include <netaddress.h>
 #include <netbase.h>
+#include <node/blockstorage.h>
 #include <node/coin.h>
 #include <node/context.h>
 #include <node/transaction.h>
@@ -52,6 +52,8 @@
 #include <memory>
 #include <optional>
 #include <utility>
+
+#include <boost/signals2/signal.hpp>
 
 using interfaces::BlockTip;
 using interfaces::Chain;
